// src/modules/doctor/domain/enums/SpecialityEnum.ts
export enum SpecialityEnum {
  CARDIOLOGY = 'CARDIOLOGY',
  DERMATOLOGY = 'DERMATOLOGY',
  NEUROLOGY = 'NEUROLOGY',
  PEDIATRICS = 'PEDIATRICS',
  ORTHOPEDICS = 'ORTHOPEDICS',
  PSYCHIATRY = 'PSYCHIATRY',
  GENERAL_MEDICINE = 'GENERAL_MEDICINE',
  ENT = 'ENT',
  OPHTHALMOLOGY = 'OPHTHALMOLOGY',
  GYNECOLOGY = 'GYNECOLOGY',
}